+++ cmd/hidden_lake/helpers/adapters/adapters_test.go
@@ -1,0 +1,5 @@
+package adapters
+
+import "testing"
+
+func TestNothing(_ *testing.T) {}